+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## <small>2.3.1 (2025-09-04)</small>
+
+- Fixed crypto.randomUUID is not available on http://192.168.x.y by adding a fallback to crypto.getRandomValues
+
 ## <small>2.3.0 (2025-09-03)</small>
 
 - Added support for multiple accounts